Time: 262 ms (69.75%) | Memory: 118.8 MB (72.54%) - LeetSync

+++ leetcode/3299-find-the-maximum-number-of-elements-in-subset/find-the-maximum-number-of-elements-in-subset.cpp
@@ -1,11 +1,12 @@
 class Solution {
 public:
+   
     int maximumLength(vector<int>& nums) {
-        long long ans=0;
+        long ans=0;
         map<int,int> freq;
         for(int n: nums) freq[n]++;
         for(auto [k,f] : freq){
-            long long t=k,count=0;
+            long t=k,count=0;
             if( t == 1)
             {
                 count+=freq[t];
